+++ kinto/plugins/admin/__init__.py
@@ -8,7 +8,7 @@ def includeme(config):
     # Expose capability.
     config.add_api_capability(
         "admin",
-        version="1.7.0",
+        version="1.8.0",
         description="Serves the admin console.",
         url="https://github.com/Kinto/kinto-admin/",
     )