+++ CHANGELOG.rst
@@ -3,14 +3,27 @@ Changelog
 
 This document describes changes between each past release.
 
-5.3.0 (unreleased)
+5.4.0 (unreleased)
 ------------------
+
+- Nothing changed yet.
+
+
+5.3.0 (2017-01-20)
+------------------
+
+**Bug fixes**
+
+- Fix crash with batch endpoint when list of requests contains trailing comma (fixes #1024)
 
 **Internal changes**
 
 - Cache backend transactions are not bound to the request/response cycle anymore (fixes #879)
 - Quick mention of PostgreSQL commands to run tests locally in contributing docs.
 - Use YAML ``safe_load`` for the swagger file. (#1022)
+- Request headers and querystrings are now validated using cornice schemas (#873).
+- JSON Patch format is now validated using cornice (#880).
+- Upgraded to Kinto-Admin 1.8.0
 
 
 5.2.0 (2017-01-11)